requirements.txt
langgraph
langchain-google-vertexai
python-dotenv
requests
pydantic
fastapi
uvicorn[standard]
